fix: SetPlayXpCap must cast daily max values to long instead of int

+++ Source/ACE.Server/Features/Xp/XpManager.cs
@@ -143,9 +143,9 @@ namespace ACE.Server.Features.Xp
             var playerTotalXp = player.GetProperty(ACE.Entity.Enum.Properties.PropertyInt64.TotalExperience);
             var diff = (long)CurrentDailyXp.XpCap - (long)playerTotalXp;
 
-            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.QuestXpDailyMax, (int)(diff * 0.4));
-            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.MonsterXpDailyMax, (int)(diff * 0.4));
-            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.PvpXpDailyMax, (int)(diff * 0.2));
+            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.QuestXpDailyMax, (long)(diff * 0.4));
+            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.MonsterXpDailyMax, (long)(diff * 0.4));
+            player.SetProperty(ACE.Entity.Enum.Properties.PropertyInt64.PvpXpDailyMax, (long)(diff * 0.2));
         }
 
         public static double? MaxLevel = null;
